+++ custom_components/dashboardmodern/frontend/legacy/build-info.js
@@ -12,4 +12,4 @@ import "../src/sections/beta24-energy-recovery-section.js";
 import "../src/sections/beta25-real-device-fixes-section.js";
 import "../src/sections/beta25-compatibility-section.js";
 import "../src/sections/beta26-real-device-stability-section.js";
-export const BUILD_INFO = Object.freeze({"generated":true,"integrationVersion":"1.3.11","dashboardVersion":"1.3.11","moduleVersion":14,"schemaVersion":4,"date":"2026-08-30T10:31:50+00:00","commit":"a17cc69c905f3cd292cb5d26654c0df028b2ad05","assetHash":"5472c6f8633355ac"});
+export const BUILD_INFO = Object.freeze({"generated":true,"integrationVersion":"1.3.12","dashboardVersion":"1.3.12","moduleVersion":14,"schemaVersion":4,"date":"2026-08-30T11:36:06+00:00","commit":"3bb34d11f4ef2d338dcafdd2cbddef0630e488c7","assetHash":"2930e731bfc989d4"});